complete UC1 implements

+++ src/main/java/it/APS/Eat/Home/demo/service/Citta/CittaServiceImpl.java
@@ -1,8 +1,11 @@
 package it.APS.Eat.Home.demo.service.Citta;
 
 import it.APS.Eat.Home.demo.Exception.NotFoundException;
+import it.APS.Eat.Home.demo.model.AcmeHome;
 import it.APS.Eat.Home.demo.model.Citta;
+import it.APS.Eat.Home.demo.model.Ristorante;
 import it.APS.Eat.Home.demo.repository.CittaRepository;
+import it.APS.Eat.Home.demo.repository.RistoranteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -13,6 +16,9 @@ public class CittaServiceImpl implements CittaService {
 
     @Autowired
     private CittaRepository cittaRepository;
+
+    @Autowired
+    private RistoranteRepository ristoranteRepository;
 
     @Override
     public List<Citta> getAllCitta() {
@@ -30,7 +36,11 @@ public class CittaServiceImpl implements CittaService {
 
     @Override
     public Citta getCittaByCodice(String codice) {
-        return cittaRepository.findById(codice).orElse(null);
+        if (this.cittaRepository.existsById(codice)) {
+            return cittaRepository.findById(codice).orElse(null);
+        } else {
+            throw new NotFoundException("La Citta non Esiste");
+        }
     }
 
     @Override
@@ -48,9 +58,30 @@ public class CittaServiceImpl implements CittaService {
 
     @Override
     public Citta deleteCittaByCodice(String codice) {
-        Citta c = this.getCittaByCodice(codice);
-        this.cittaRepository.delete(c);
-        return c;
+        if (this.cittaRepository.existsById(codice)) {
+            Citta c = this.getCittaByCodice(codice);
+            this.cittaRepository.delete(c);
+            return c;
+        } else {
+            throw new NotFoundException("La Citta non Esiste");
+        }
+    }
+
+    @Override
+    public Citta confermaInserimentoRistorante(String codiceRistorante) {
+        if (this.ristoranteRepository.existsById(codiceRistorante)) {
+            Ristorante ristorante = ristoranteRepository.findById(codiceRistorante).orElse(null);
+            if (this.cittaRepository.existsById(ristorante.getCodiceCitta())) {
+                Citta c = this.cittaRepository.findById(ristorante.getCodiceCitta()).orElse(null);
+                c.getRistoranti().add(ristorante);
+                this.updateCitta(c, c.getCodiceCitta());
+                return c;
+            } else {
+                throw new NotFoundException("La Citta non Esiste");
+            }
+        } else {
+            throw new NotFoundException("Il Ristorante non Esiste");
+        }
     }
 
 }
